the quote is now changing when pressing the button
